Create custom global preferences configuration fragment
PreferenceFragment is tightly coupled with SharedPreferences, so making it work
with any other ways to store data is painful. It also enforces a specific
behavior (such as automatic saving on each change) and lacks configuration,
making it unsuitable if you don't like that behavior.

As I'm planning to change how global preferences are represented in the code,
PreferenceFragment is not suitable for this case.