Move default configuration values from resource XMLs to GlobalPreferences
This way we gain no-arg GlobalPreferences constructor (as opposed to
constructor requiring values for all properties), defining defaults is simpler,
and it's easier to find them. The main cost is that view templates do not have
these defaults applied, but user settings are applied to them on activity
creation, so in the end it doesn't matter.